Update config.env to include REPO env variable
Introduced REPO environment variable that the cli also understands. It is embedded in the pre-existing EDGE_REPO variable defined in the config.env to avoid any test case changes. This change should allow a seamless change of the devel/upstream/download repo without requiring additional export commands on the shell.